requirements.txt
mysql-connector-python
colorama
ecdsa
pyfiglet